+++ include/tao/pegtl/internal/must.hpp
@@ -53,7 +53,7 @@ namespace TAO_PEGTL_NAMESPACE::internal
                 typename... States >
       [[nodiscard]] static bool match( ParseInput& in, States&&... st )
       {
-         if( !Control< Rule >::template match< A, rewind_mode::dontcare, Action, Control >( in, st... ) ) {
+         if( !Control< Rule >::template match< A, rewind_mode::optional, Action, Control >( in, st... ) ) {
             Control< Rule >::raise( static_cast< const ParseInput& >( in ), st... );
          }
          return true;